requirements.txt
playwright==1.44.0
python-dotenv==1.0.0
